refactor: change monetary values from decimal to integer

+++ database/migrations/2026_04_28_195929_create_accounts_table.php
@@ -16,7 +16,7 @@ return new class () extends Migration {
             $table->id();
             $table->foreignId('user_id')->constrained()->restrictOnDelete();
             $table->string('name', 100);
-            $table->decimal('balance', 15, 2);
+            $table->bigInteger('balance');
             $table->timestamps();
         });
     }
